requirements.txt
aniso8601==7.0.0
Click==7.0
Flask==1.1.1
Flask-HTTPAuth==3.3.0
Flask-RESTful==0.3.7
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
pymongo==3.9.0
python-dotenv==0.10.3
pytz==2019.2
six==1.12.0
Werkzeug==0.15.5
